refatorado: combinando as classes para simplificar
